refactor: extract transcript viewport
